+++ src/constants/languages.js
@@ -1,0 +1,9 @@
+// Constants will be shared between client and server.
+// They will be used as values in the DOM and in the URL of the referrer calls.
+const LANGUAGES_SORTING_TOP = 'top'
+const LANGUAGES_SORTING_RECENT = 'recent'
+
+module.exports = {
+	LANGUAGES_SORTING_TOP,
+	LANGUAGES_SORTING_RECENT
+}