+++ src/server.js
@@ -14,7 +14,9 @@ const tokens = require('./routes/tokens')
 const domains = require('./routes/domains')
 const records = require('./routes/records')
 const views = require('./routes/views')
+const pages = require('./routes/pages')
 const referrers = require('./routes/referrers')
+const languages = require('./routes/languages')
 
 const catchError = (fn) => async (req, res) => {
 
@@ -24,18 +26,27 @@ const catchError = (fn) => async (req, res) => {
 
 	} catch (err) {
 
-		signale.fatal(err)
+		const isUnknownError = err.statusCode == null
+		const hasOriginalError = err.originalError != null
 
-		if (err.statusCode != null) send(res, err.statusCode, err.message)
-		else send(res, 500, err.message)
+		// Only log the full error stack when the error isn't a known API response
+		if (isUnknownError === true) {
+			signale.fatal(err)
+			return send(res, 500, err.message)
+		}
+
+		signale.warn(hasOriginalError === true ? err.originalError.message : err.message)
+		send(res, err.statusCode, err.message)
 
 	}
 
 }
 
-const notFound = async () => {
+const notFound = async (req) => {
 
-	throw createError(404, 'Not found')
+	const err = new Error(`\`${ req.url }\` not found`)
+
+	throw createError(404, 'Not found', err)
 
 }
 
@@ -63,7 +74,11 @@ module.exports = micro(
 
 			get('/domains/:domainId/views', pipe(requireAuth, views.get)),
 
+			get('/domains/:domainId/pages', pipe(requireAuth, pages.get)),
+
 			get('/domains/:domainId/referrers', pipe(requireAuth, referrers.get)),
+
+			get('/domains/:domainId/languages', pipe(requireAuth, languages.get)),
 
 			get('/*', notFound),
 			post('/*', notFound),