+++ src/ui/scripts/components/routes/RouteLanguages.js
@@ -1,0 +1,60 @@
+import { createElement as h, Fragment, useEffect } from 'react'
+
+import {
+	LANGUAGES_SORTING_TOP,
+	LANGUAGES_SORTING_RECENT
+} from '../../../../constants/languages'
+
+import enhanceLanguages from '../../utils/enhanceLanguages'
+
+import CardLanguages from '../cards/CardLanguages'
+import Select from '../Select'
+
+const RouteLanguages = (props) => {
+
+	useEffect(() => {
+
+		props.fetchDomains(props)
+
+	}, [])
+
+	useEffect(() => {
+
+		props.domains.value.map((domain) => {
+			props.fetchLanguages(props, domain.data.id)
+		})
+
+	}, [ props.domains.value, props.languages.sorting ])
+
+	return (
+		h(Fragment, {},
+
+			h('div', { className: 'subHeader' },
+				h(Select, {
+					value: props.languages.sorting,
+					onChange: (e) => props.setLanguagesSorting(e.target.value),
+					items: [
+						{ value: LANGUAGES_SORTING_TOP, label: 'Top languages' },
+						{ value: LANGUAGES_SORTING_RECENT, label: 'Recent languages' }
+					]
+				})
+			),
+
+			props.domains.value.map(
+				(domain) => (
+					h(CardLanguages, {
+						key: domain.data.id,
+						headline: domain.data.title,
+						sorting: props.languages.sorting,
+						loading: props.languages.value[domain.data.id] == null ? false : props.languages.value[domain.data.id].fetching,
+						items: props.languages.value[domain.data.id] == null ? [] : enhanceLanguages(props.languages.value[domain.data.id].value)
+					})
+				)
+			)
+
+		)
+	)
+
+}
+
+export default RouteLanguages