+++ src/ui/scripts/index.js
@@ -14,7 +14,9 @@ import * as actions from './actions'
 import { initialState as initialTokenState } from './reducers/token'
 import { initialState as initialRouteState } from './reducers/route'
 import { initialState as initialViewsState } from './reducers/views'
+import { initialState as initialPagesState } from './reducers/pages'
 import { initialState as initialReferrersState } from './reducers/referrers'
+import { initialState as initialLanguagesState } from './reducers/languages'
 
 import Main from './components/Main'
 
@@ -48,9 +50,17 @@ store.subscribe(() => {
 			...initialViewsState(),
 			type: currentState.views.type
 		},
+		pages: {
+			...initialPagesState(),
+			type: currentState.pages.type
+		},
 		referrers: {
 			...initialReferrersState(),
 			sorting: currentState.referrers.sorting
+		},
+		languages: {
+			...initialLanguagesState(),
+			sorting: currentState.languages.sorting
 		}
 	})