+++ src/ui/scripts/reducers/index.js
@@ -5,7 +5,9 @@ import token from './token'
 import route from './route'
 import domains from './domains'
 import views from './views'
+import pages from './pages'
 import referrers from './referrers'
+import languages from './languages'
 
 export default combineReducers({
 	modals,
@@ -13,5 +15,7 @@ export default combineReducers({
 	route,
 	domains,
 	views,
-	referrers
+	pages,
+	referrers,
+	languages
 })