+++ src/utils/normalizeUrl.js
@@ -1,0 +1,13 @@
+'use strict'
+
+const normalizeUrl = require('normalize-url')
+
+module.exports = (url) => normalizeUrl(url, {
+	removeDirectoryIndex: true,
+	removeQueryParameters: [
+		/^utm_\w+/i,
+		'ref',
+		'fbclid',
+		'source'
+	]
+})