+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Top and recent languages
+- Comparison bars for "Views per page"
+
+### Changed
+
+- Top and recent pages in a dedicated view
+- Hover recent referrers to see the date
+- Improved URL normalization for cleaner URLs in "Views per page" and "Referrers"
+- Improved logging of errors in the server log
+- Improved menu on small screens
+- Reset state after pressing "Reload Ackee" in the error overlay to recover from bugs caused by a faulty state
+
 ## [1.1.0] - 2019-09-11
 
 > ⚠️ All options / environment variables have been renamed. They're now starting with `ACKEE_` to avoid collisions with other tools. Please update your options accordingly.